Dedupe split_pem_bundle
